+++ src/Neo.L2.Abstractions/Models/NeoHubDeployReport.cs
@@ -446,7 +446,7 @@ public sealed record NeoHubDeployReport(
                         + "IsDepositPipelineWiringComplete / IsMessagePipelineWiringComplete / "
                         + "IsForcedInclusionPipelineWiringComplete / IsSettlementClientWiringComplete / "
                         + "IsPipelineEnabled / IsSettlementPoisoned / IsSettlementIdle / IsPipelineHealthy / "
-                        + "PipelineHealthFailures / IsMetricsHttpHealthy / MetricsHttpHealthFailures / "
+                        + "PipelineHealthFailures / HasOverdueForcedInclusion / IsMetricsHttpHealthy / MetricsHttpHealthFailures / "
                         + "IsLocalHostHealthy / LocalHostHealthFailures / "
                         + "HasExpectedNetwork / HasScannerDeployHeights / IsOfflinePassportComplete / "
                         + "OfflinePassportFailures / "